conda-requirements.txt
snakemake=3.13.0
ruamel.yaml
biopython
trimmomatic
fastqc
blast
CLARK
kraken=1.0
krona
kraken-biom
pear
vsearch
bwa
bowtie2
jellyfish=1.1.11=1
cutadapt
pysam
pandas
igv=2.3.98
megahit
cap3
prodigal
setuptools_scm
